+++ src/extension.ts
@@ -149,7 +149,8 @@ export async function activate(context: vscode.ExtensionContext) {
     subscriptions.push(vscode.commands.registerCommand('eide.project.uploadToDevice', (item) => projectExplorer.UploadToDevice(item)));
     subscriptions.push(vscode.commands.registerCommand('eide.reinstall.binaries', () => checkAndInstallBinaries(true)));
     subscriptions.push(vscode.commands.registerCommand('eide.project.flash.erase.all', (item) => projectExplorer.UploadToDevice(item, true)));
-    subscriptions.push(vscode.commands.registerCommand('eide.project.buildAndFlash', (item) => projectExplorer.BuildSolution(item, { useFastMode: true }, true)));
+    subscriptions.push(vscode.commands.registerCommand('eide.project.buildAndFlash', (item) => projectExplorer.BuildSolution(item, { useFastMode: true, flashAfterBuild: true })));
+    subscriptions.push(vscode.commands.registerCommand('eide.project.genBuilderParams', (item) => projectExplorer.BuildSolution(item, { useFastMode: true, onlyGenParams: true })));
 
     // operations bar
     subscriptions.push(vscode.commands.registerCommand('_cl.eide.project.historyRecord', () => projectExplorer.openHistoryRecords()));
@@ -726,86 +727,114 @@ function exportEnvToSysPath() {
 
 async function checkAndInstallRuntime() {
 
+    const dotnet_chk_cmd = `dotnet --list-runtimes`;
+
     //
     // if not found dotnet, preset dotnet root folder into system env
     // dotnet path: 'C:\Program Files\dotnet'
     //
     if (os.platform() == 'win32') {
         try {
-            ChildProcess.execSync(`dotnet --info`);
+            ChildProcess.execSync(dotnet_chk_cmd);
         } catch (error) {
-            platform.appendToSysEnv(process.env, ['C:\\Program Files\\dotnet']);
+            platform.appendToSysEnv(process.env, ['C:\\Program Files\\dotnet']);        // for win x64
+            platform.appendToSysEnv(process.env, ['C:\\Program Files (x86)\\dotnet']);  // for win x86
         }
     }
 
     //
-    // check or install .NET
+    // check/install .NET
     //
     try {
-        GlobalEvent.emit('globalLog', newMessage('Info', 'Checking .NET6 Runtime ...'));
-        const chkCmd = `dotnet --info`;
-        const dotnetInfo = ChildProcess.execSync(chkCmd).toString().trim();
-        GlobalEvent.emit('globalLog', newMessage('Info', `Exec '${chkCmd}' :\n${dotnetInfo}`));
+        GlobalEvent.emit('globalLog', newMessage('Info', 'Checking .NET6 runtime ...'));
+        const dotnetInfo = ChildProcess.execSync(dotnet_chk_cmd).toString().trim();
+        GlobalEvent.emit('globalLog', newMessage('Info', `Exec cmd: '${dotnet_chk_cmd}'\n${dotnetInfo}`));
         // check dotnet version
-        if (/Microsoft\.NETCore\.App 6\./.test(dotnetInfo) ||
-            /\.NET runtimes installed:/.test(dotnetInfo)) {
-            GlobalEvent.emit('globalLog', newMessage('Info', '.NET6 Runtime Found !'));
-        } else {
-            throw new Error(`Not found .NET6 Runtime`);
+        let dotnetVerLine: string | undefined;
+        const lines = dotnetInfo.trim().split(/\r\n|\n/);
+        for (const line_ of lines) {
+            const line = line_.trim();
+            if (line.startsWith('Microsoft.NETCore.App 6.')) {
+                dotnetVerLine = line;
+                GlobalEvent.emit('globalLog', newMessage('Info', `.NET6 runtime: '${dotnetVerLine}' found !`));
+                break;
+            }
         }
+        if (!dotnetVerLine) { throw new Error(`Not found .NET6 runtime`); }
     } catch (error) {
+
+        GlobalEvent.emit('globalLog.show'); // show error log for user
 
         GlobalEvent.emit('globalLog', newMessage('Info', 'Not found [.NET6 Runtime](https://dotnet.microsoft.com/en-us/download/dotnet/6.0) !'));
 
+        /* @deprecated
         const msg = `Not found [.NET6 Runtime](https://dotnet.microsoft.com/en-us/download/dotnet/6.0), please install it !`;
         const sel = await vscode.window.showWarningMessage(msg, txt_install_now);
-        if (!sel) { return } // user canceled
+        if (!sel) { return } // user canceled */
 
         // for other platform, user need install it manually
         if (os.platform() != 'win32') {
+            const msg = `Not found [.NET6 runtime](https://dotnet.microsoft.com/en-us/download/dotnet/6.0) on your pc, please install it !`;
+            vscode.window.showWarningMessage(msg);
             // https://dotnet.microsoft.com/en-us/download/dotnet/scripts
             utility.openUrl(`https://dotnet.microsoft.com/en-us/download/dotnet/scripts`);
         }
 
         // win32, we can auto install it
         else {
-            const tmpFile = File.fromArray([os.tmpdir(), 'dotnet-runtime-6.0.5-win-x64.exe']);
+            const defPkgName = 'dotnet-runtime-6.0.5-win-x64.exe';
 
-            try {
-                if (tmpFile.IsFile()) { fs.unlinkSync(tmpFile.path) }
-            } catch (error) {
-                // do nothing
+            let pkgReady: boolean;
+            let pkgFile: File;
+
+            // if found local installer pkg, use it
+            const localPkg = File.fromArray([ResManager.GetInstance().getAppRootFolder().path, defPkgName]);
+            if (localPkg.IsFile()) {
+                pkgReady = true;
+                pkgFile = localPkg;
             }
 
-            const downloadUrl = `https://download.visualstudio.microsoft.com/download/pr/b395fa18-c53b-4f7f-bf91-6b2d3c43fedb/d83a318111da9e15f5ecebfd2d190e89/dotnet-runtime-6.0.5-win-x64.exe`;
+            // if not found local installer pkg, download it
+            else {
 
-            const done = await vscode.window.withProgress({
-                location: vscode.ProgressLocation.Notification,
-                title: 'Downloading .NET6 installer',
-                cancellable: false
-            }, async (progress, token): Promise<boolean> => {
+                pkgFile = File.fromArray([os.tmpdir(), defPkgName]);
 
-                const res = await utility.downloadFileWithProgress(downloadUrl, tmpFile.name, progress, token);
-
-                if (res instanceof Buffer) {
-                    try {
-                        fs.writeFileSync(tmpFile.path, res);
-                        return true;
-                    } catch (error) {
-                        return false;
-                    }
-                }
-
-                if (res instanceof Error) {
-                    GlobalEvent.emit('error', res);
-                }
-
-                return false;
-            });
-
-            if (done && tmpFile.IsFile()) {
                 try {
-                    ChildProcess.execFileSync(tmpFile.path);
+                    if (pkgFile.IsFile()) { fs.unlinkSync(pkgFile.path) }
+                } catch (error) {
+                    // do nothing
+                }
+
+                const downloadUrl = `https://download.visualstudio.microsoft.com/download/pr/b395fa18-c53b-4f7f-bf91-6b2d3c43fedb/d83a318111da9e15f5ecebfd2d190e89/dotnet-runtime-6.0.5-win-x64.exe`;
+
+                pkgReady = await vscode.window.withProgress({
+                    location: vscode.ProgressLocation.Notification,
+                    title: 'Downloading .NET6 runtime installer',
+                    cancellable: false
+                }, async (progress, token): Promise<boolean> => {
+
+                    const res = await utility.downloadFileWithProgress(downloadUrl, pkgFile.name, progress, token);
+
+                    if (res instanceof Buffer) {
+                        try {
+                            fs.writeFileSync(pkgFile.path, res);
+                            return true;
+                        } catch (error) {
+                            return false;
+                        }
+                    }
+
+                    if (res instanceof Error) {
+                        GlobalEvent.emit('error', res);
+                    }
+
+                    return false;
+                });
+            }
+
+            if (pkgReady && pkgFile.IsFile()) {
+                try {
+                    ChildProcess.execFileSync(pkgFile.path);
                     const sel = await vscode.window.showInformationMessage(`Ok ! Now you need relaunch VsCode !`, txt_yes);
                     if (sel) {
                         vscode.commands.executeCommand('workbench.action.reloadWindow');
@@ -814,7 +843,7 @@ async function checkAndInstallRuntime() {
                     GlobalEvent.emit('msg', newMessage('Error', `Install [.NET6 runtime](https://dotnet.microsoft.com/en-us/download/dotnet/6.0) failed, you need install it manually !`));
                 }
             } else {
-                vscode.window.showWarningMessage(`Install .NET6 failed, you need install it manually !`);
+                vscode.window.showWarningMessage(`Install .NET6 runtime failed, you need install it manually !`);
                 // https://dotnet.microsoft.com/en-us/download/dotnet/scripts
                 // https://dotnet.microsoft.com/en-us/download/dotnet/thank-you/runtime-6.0.5-windows-x64-installer
                 utility.openUrl(`https://dotnet.microsoft.com/en-us/download/dotnet/thank-you/runtime-6.0.5-windows-x64-installer`);
@@ -1019,8 +1048,8 @@ function RegisterGlobalEvent() {
 
     const outChannel = vscode.window.createOutputChannel('eide-log');
     GlobalEvent.on('globalLog', (msg) => outChannel.appendLine(LogDumper.Msg2String(msg)));
-    GlobalEvent.on('eide.log.append', (log) => outChannel.append(log));
-    GlobalEvent.on('eide.log.show', () => outChannel.show());
+    GlobalEvent.on('globalLog.append', (log) => outChannel.append(log));
+    GlobalEvent.on('globalLog.show', () => outChannel.show());
 
     GlobalEvent.on('project.opened', () => {
         prj_count++; // increment cnt